+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/handlers/ItemUseHandler.java
@@ -1,0 +1,37 @@
+package io.github.thebusybiscuit.slimefun4.core.handlers;
+
+import org.bukkit.entity.Player;
+
+import io.github.thebusybiscuit.slimefun4.api.events.PlayerRightClickEvent;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
+import me.mrCookieSlime.Slimefun.Objects.handlers.ItemHandler;
+
+/**
+ * This {@link ItemHandler} is triggered when the {@link SlimefunItem} it was assigned to
+ * is right-clicked.
+ * 
+ * @author TheBusyBiscuit
+ *
+ * @see ItemHandler
+ * @see SimpleSlimefunItem
+ * 
+ */
+@FunctionalInterface
+public interface ItemUseHandler extends ItemHandler {
+
+    /**
+     * This function is triggered when a {@link Player} right clicks with the assigned {@link SlimefunItem}
+     * in his hand.
+     * 
+     * @param e
+     *            The {@link PlayerRightClickEvent} that was triggered
+     */
+    void onRightClick(PlayerRightClickEvent e);
+
+    @Override
+    default Class<? extends ItemHandler> getIdentifier() {
+        return ItemUseHandler.class;
+    }
+
+}
