+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/seasonal/ChristmasPresent.java
@@ -7,11 +7,11 @@ import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.inventory.ItemUtils;
 import io.github.thebusybiscuit.slimefun4.core.attributes.NotPlaceable;
+import io.github.thebusybiscuit.slimefun4.core.handlers.BlockPlaceHandler;
 import io.github.thebusybiscuit.slimefun4.utils.FireworkUtils;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
-import me.mrCookieSlime.Slimefun.Objects.handlers.BlockPlaceHandler;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
 public class ChristmasPresent extends SimpleSlimefunItem<BlockPlaceHandler> implements NotPlaceable {
@@ -26,7 +26,7 @@ public class ChristmasPresent extends SimpleSlimefunItem<BlockPlaceHandler> impl
 
     @Override
     public BlockPlaceHandler getItemHandler() {
-        return (e, item) -> {
+        return (p, e, item) -> {
             if (isItem(item)) {
                 e.setCancelled(true);
 
@@ -34,7 +34,7 @@ public class ChristmasPresent extends SimpleSlimefunItem<BlockPlaceHandler> impl
                     ItemUtils.consumeItem(item, false);
                 }
 
-                FireworkUtils.launchRandom(e.getPlayer(), 3);
+                FireworkUtils.launchRandom(p, 3);
 
                 ItemStack gift = gifts[ThreadLocalRandom.current().nextInt(gifts.length)].clone();
                 e.getBlockPlaced().getWorld().dropItemNaturally(e.getBlockPlaced().getLocation(), gift);
