+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/SlimefunItem/SlimefunItem.java
@@ -391,6 +391,7 @@ public class SlimefunItem implements Placeable {
                 }
 
                 state = ItemState.ENABLED;
+                checkForDeprecations(getClass());
 
                 useableInWorkbench = SlimefunPlugin.getItemCfg().getBoolean(id + ".can-be-used-in-workbenches");
                 hidden = SlimefunPlugin.getItemCfg().getBoolean(id + ".hide-in-guide");
@@ -426,11 +427,46 @@ public class SlimefunItem implements Placeable {
             if (exception.isPresent()) {
                 throw exception.get();
             }
+            else {
+                // Make developers or at least Server admins aware that
+                // an Item is using a deprecated ItemHandler
+                // checkForDeprecations(handler.getClass());
+                // A bit too spammy atm, will enable it again later
+            }
 
             if (!handler.isPrivate()) {
                 Set<ItemHandler> handlerset = getPublicItemHandlers(handler.getIdentifier());
                 handlerset.add(handler);
             }
+        }
+    }
+
+    /**
+     * This method checks recursively for all {@link Class} parents to look for any {@link Deprecated}
+     * elements.
+     * 
+     * If a {@link Deprecated} element was found, a warning message will be printed.
+     * 
+     * @param c
+     *            The {@link Class} from which to start this operation.
+     */
+    private void checkForDeprecations(Class<?> c) {
+        // We do not wanna throw an Exception here since this could also mean that
+        // we have reached the end of the Class hierarchy
+        if (c != null) {
+            // Check if this Class is deprecated
+            if (c.isAnnotationPresent(Deprecated.class)) {
+                warn("The inherited Class \"" + c.getName() + "\" has been deprecated. Check the documentation for more details!");
+            }
+
+            for (Class<?> parent : c.getInterfaces()) {
+                // Check if this Interface is deprecated
+                if (parent.isAnnotationPresent(Deprecated.class)) {
+                    warn("The implemented Interface \"" + parent.getName() + "\" has been deprecated. Check the documentation for more details!");
+                }
+            }
+
+            checkForDeprecations(c.getSuperclass());
         }
     }
 
@@ -765,6 +801,11 @@ public class SlimefunItem implements Placeable {
     public void warn(String message) {
         String msg = toString() + ": " + message;
         addon.getLogger().log(Level.WARNING, msg);
+
+        if (addon.getBugTrackerURL() != null) {
+            // We can prompt the server operator to report it to the addon's bug tracker
+            addon.getLogger().log(Level.WARNING, "You can report this warning here: {0}", addon.getBugTrackerURL());
+        }
     }
 
     /**
