+++ src/main/java/me/mrCookieSlime/Slimefun/Objects/handlers/ItemUseHandler.java
@@ -1,36 +1,9 @@
 package me.mrCookieSlime.Slimefun.Objects.handlers;
 
-import org.bukkit.entity.Player;
-
-import io.github.thebusybiscuit.slimefun4.api.events.PlayerRightClickEvent;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SimpleSlimefunItem;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
-
 /**
- * This {@link ItemHandler} is triggered when the {@link SlimefunItem} it was assigned to
- * is right-clicked.
- * 
- * @author TheBusyBiscuit
- *
- * @see ItemHandler
- * @see SimpleSlimefunItem
- * 
+ * @deprecated Moved to {@link io.github.thebusybiscuit.slimefun4.core.handlers.ItemUseHandler}
  */
-@FunctionalInterface
-public interface ItemUseHandler extends ItemHandler {
-
-    /**
-     * This function is triggered when a {@link Player} right clicks with the assigned {@link SlimefunItem}
-     * in his hand.
-     * 
-     * @param e
-     *            The {@link PlayerRightClickEvent} that was triggered
-     */
-    void onRightClick(PlayerRightClickEvent e);
-
-    @Override
-    default Class<? extends ItemHandler> getIdentifier() {
-        return ItemUseHandler.class;
-    }
+@Deprecated
+public interface ItemUseHandler extends io.github.thebusybiscuit.slimefun4.core.handlers.ItemUseHandler {
 
 }
